+++ suite-common/react-query/react/index.ts
@@ -1,0 +1,1 @@
+export { ReactQueryProvider } from '../src/components/ReactQueryProvider';
